backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlmodel==0.0.16
pydantic-settings==2.2.1
python-dotenv==1.0.1
